Fix code for FileUtils class
