+++ pravega-sensor-collector/src/main/java/io/pravega/sensor/collector/parquet/ParquetFileProcessor.java
@@ -143,39 +143,6 @@ public class ParquetFileProcessor {
     }
 
     /**
-     * @return list of file name and file size in bytes
-     */
-    static protected List<FileNameWithOffset> getDirectoryListing(String fileSpec, String fileExtension) throws IOException {
-        final Path pathSpec = Paths.get(fileSpec);
-        if (!Files.isDirectory(pathSpec.toAbsolutePath())) {
-            log.error("getDirectoryListing: Directory does not exist or spec is not valid : {}", pathSpec.toAbsolutePath());
-            throw new IOException("Directory does not exist or spec is not valid");
-        }
-        List<FileNameWithOffset> directoryListing = new ArrayList<>();
-        try(DirectoryStream<Path> dirStream=Files.newDirectoryStream(pathSpec)){
-            for(Path path: dirStream){
-                if(Files.isDirectory(path))         //traverse subdirectories
-                    directoryListing.addAll(getDirectoryListing(path.toString(), fileExtension));
-                else {
-                    FileNameWithOffset fileEntry = new FileNameWithOffset(path.toAbsolutePath().toString(), path.toFile().length());
-                    if(isValidFile(fileEntry, fileExtension)){
-                        directoryListing.add(fileEntry);
-                    }
-                }
-            }
-        }catch(Exception ex){
-            if(ex instanceof IOException){
-                log.error("getDirectoryListing: Directory does not exist or spec is not valid : {}", pathSpec.toAbsolutePath());
-                throw new IOException("Directory does not exist or spec is not valid");
-            }else{
-                log.error("getDirectoryListing: Exception while listing files: {}", pathSpec.toAbsolutePath());
-                throw new IOException(ex);
-            }
-        }
-        return directoryListing;
-    }
-
-    /**
      * @return sorted list of file name and file size in bytes
      */
     static protected List<FileNameWithOffset> getNewFiles(List<FileNameWithOffset> directoryListing, List<FileNameWithOffset> completedFiles) {
@@ -288,25 +255,6 @@ public class ParquetFileProcessor {
                 // We can continue on this error. Deletion will be retried on the next iteration.
             }
         });
-    }
-
-    /*
-    Check for below file validation
-        1. Is File empty
-        2. If extension is null or extension is valid ingest all file
-     */
-    public static boolean isValidFile(FileNameWithOffset fileEntry, String fileExtension ){
-
-            if(fileEntry.offset<=0){
-                log.warn("isValidFile: Empty file {} can not be processed ",fileEntry.fileName);
-            }
-            // If extension is null, ingest all files
-            else if(fileExtension.isEmpty() || fileExtension.equals(fileEntry.fileName.substring(fileEntry.fileName.lastIndexOf(".")+1)))
-                return true;
-            else
-                log.warn("isValidFile: File format {} is not supported ", fileEntry.fileName);
-
-        return false;
     }
 
     /**
