+++ pravega-sensor-collector/src/test/java/io/pravega/sensor/collector/file/LogFileSequenceProcessorStateTests.java
@@ -17,6 +17,8 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import io.pravega.sensor.collector.util.FileNameWithOffset;
+
 import java.sql.SQLException;
 import java.util.HashSet;
 import java.util.List;
