+++ pravega-sensor-collector/src/test/java/io/pravega/sensor/collector/file/LogFileSequenceProcessorTests.java
@@ -15,6 +15,9 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import io.pravega.sensor.collector.util.FileNameWithOffset;
+import io.pravega.sensor.collector.util.FileUtils;
+
 import java.io.IOException;
 import java.util.List;
 
@@ -39,7 +42,7 @@ public class LogFileSequenceProcessorTests {
 
     @Test
     public void getDirectoryListingTest() throws IOException {
-        final List<FileNameWithOffset> actual = LogFileSequenceProcessor.getDirectoryListing(
+        final List<FileNameWithOffset> actual = FileUtils.getDirectoryListing(
                 "../log-file-sample-data/","csv");
         log.info("actual={}", actual);
     }
